M Series Macs/requirements.txt
cmake
dlib==19.24.2    
face_recognition   
setuptools
numpy==1.26.4
opencv-python
openai-whisper
elevenlabs
sounddevice
soundfile
SpeechRecognition
pyobjc
webrtcvad
geocoder
langchain-community
faiss_cpu
token
pyaudio
python-dotenv
pygame
